backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
pikepdf==8.10.1
Pillow==10.2.0
python-dotenv==1.0.0
azure-storage-blob==12.19.0
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
pdf2docx==0.5.8
pdf2image==1.17.0
python-docx==1.1.0
reportlab==4.0.8
openpyxl==3.1.2
tabula-py==2.9.0
pandas==2.1.4
